requirements.txt
discord.py
praw
instagrapi
tweepy[async]
ShopifyAPI
